Medical-Chatbot/requirements.txt
# requirements.txt
# **Agents**
google-genai
huggingface_hub
# **RAG**
faiss-cpu
sentence-transformers
# **NLPs**
transformers
accelerate  
sentencepiece
# **Environment**
python-dotenv       # Not used in Streamlit deployment
pymongo
# **Deployment**
uvicorn
fastapi
torch               # Reduce model load with half-precision (float16) to reduce RAM usage
psutil              # CPU/RAM logger
